backend/requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
google-cloud-firestore>=2.19.0
google-cloud-aiplatform>=1.74.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
uvloop>=0.21.0
pydantic>=2.0.0
httpx>=0.27.0
